Store bytes, and move source text creation
